Abort device controller tasks if connection fails
The two subtasks in the device controller hold onto state about the
device that can interfere with reconnecting to the device later on.

Bug #2 was caused by this. I was unaware that tokio tasks are not
aborted if their JoinHandle is dropped. However, they *are* aborted if
they're part of a JoinSet which is dropped. So use a JoinSet to ensure
they get dropped - then all the relevant state is also dropped, and the
device can reconnect on the next pass through the controller main loop.

I considered using cancellation tokens to deal with this, but I think I
actually prefer the idea that the task is just aborted - they don't have
particular memory/resource leak risks, and cancellation tokens would
make the logic more complex.

Fixes #2