Extract a DeviceTemperature type, enable F display
The aim was to remove any type confusion from the rest of the code. For
example, previously there were `u16` variables lying around - are these
in wire format ("bcdish") or are they just whole degrees...?

Maybe extracting a whole type was a little much, but it allows more use
of `from` and `into`.

Secondly, the cursive UI will now display temperatures in Fahrenheit if
appropriate. It will not, however, use degrees F in the "set profile"
dialog.